python/instruct_pen/requirements.txt
numpy
pen-simulator
torch
diffusers
datasets
accelerate
peft
torchvision
tqdm
transformers
matplotlib
svgpath2mpl
pillow
opencv-python
